agents/customer_support/requirements.txt
strands-agents>=0.1.0
strands-agents-tools>=0.1.0
strands-agents-builder-bedrock-agentcore>=0.1.0
bedrock-agentcore-runtime>=0.1.0
boto3>=1.35.0
